Created 2 Classes
INTERACT: Any interactable can use this script. This would include anything like an item (which I made), a door, a chest, a shop,  etc. Anything the player needs to be able to interact with, make it a child of this class.

ITEM: Derives from the Interact class. Executes a function when interacted with. Currently has 2 other variables: description and value.