BaseItem created
made a baseitem prefab that has 2 GameAction scripts on it:

AddItem_GA: adds an item to the inventory of the player. Currently still being worked on, need to communicate with Connor about it.

UpdateStat_GA: updates one of the stats on the playerframe by a given modifier. NOTE: if you add a stat to the CharacterFrame, be sure to add it to the enum associated with it.